refactor(checker): keep globalThis access helpers under guard ceiling
