+++ crates/tsz-checker/src/state/type_resolution/core.rs
@@ -1983,23 +1983,4 @@ impl<'a> CheckerState<'a> {
         }
         TypeId::ERROR
     }
-
-    /// Resolve a primitive keyword like `number`, `string`, etc.
-    fn resolve_primitive_keyword(name: &str) -> Option<TypeId> {
-        match name {
-            "number" => Some(TypeId::NUMBER),
-            "string" => Some(TypeId::STRING),
-            "boolean" => Some(TypeId::BOOLEAN),
-            "void" => Some(TypeId::VOID),
-            "any" => Some(TypeId::ANY),
-            "never" => Some(TypeId::NEVER),
-            "unknown" => Some(TypeId::UNKNOWN),
-            "undefined" => Some(TypeId::UNDEFINED),
-            "null" => Some(TypeId::NULL),
-            "object" => Some(TypeId::OBJECT),
-            "bigint" => Some(TypeId::BIGINT),
-            "symbol" => Some(TypeId::SYMBOL),
-            _ => None,
-        }
-    }
 }
