+++ crates/tsz-checker/src/state/type_resolution/primitive_keyword.rs
@@ -1,0 +1,22 @@
+use crate::state::CheckerState;
+use tsz_solver::TypeId;
+
+impl<'a> CheckerState<'a> {
+    pub(super) fn resolve_primitive_keyword(name: &str) -> Option<TypeId> {
+        match name {
+            "number" => Some(TypeId::NUMBER),
+            "string" => Some(TypeId::STRING),
+            "boolean" => Some(TypeId::BOOLEAN),
+            "void" => Some(TypeId::VOID),
+            "any" => Some(TypeId::ANY),
+            "never" => Some(TypeId::NEVER),
+            "unknown" => Some(TypeId::UNKNOWN),
+            "undefined" => Some(TypeId::UNDEFINED),
+            "null" => Some(TypeId::NULL),
+            "object" => Some(TypeId::OBJECT),
+            "bigint" => Some(TypeId::BIGINT),
+            "symbol" => Some(TypeId::SYMBOL),
+            _ => None,
+        }
+    }
+}
